auth: Add afsconf_GetIdentity()
The afsconf_GetNthIdentity() and afsconf_GetNthUser() are implemented
with an internal function called GetNthIdentity(), which also duplicates
much of the code in afsconf_IsSuperIdentity().

Replace GetNthIdentity() a new general purpose lookup function called
GetIdentity(). This new function generalizes the lookup by scanning the
UserList and calling a comparison function on each entry to find a match
based on the caller's requirements.  Use this new function to also
remove the code dulication in afsconf_IsSuperIdentity().

For the GetNth* functions, the comparison functions count the entries
and matches on the target index number. For the IsSuperIdentity()
function, the comparison function finds a matching indentity using the
rx_identity_match() comparison.

Add and export a public function called afsconf_GetIdentity() which can
be used in the future to replace afsconf_GetNthUser() in the bosserver.

Change-Id: I543371a409d9c114cab3af4928269bd9dbd5ae97