bozo: Use afsconf_GetIdentity() to list superusers
Implement the ListSUser RPC with the new afsconf_GetIdentity() function
to lookup entries in the UserList.  This new function returns an
rx_identity so the RPC output string can be set by duplicating the
identity displayName, instead of passing a fixed sized buffer to be
filled (which could possibly be truncated).

Change-Id: I9232efffabc6e9e8c1ab5bcacad9197b06b758ed